% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airlabs_api_function.R
\name{airlabs_api_function}
\alias{airlabs_api_function}
\title{AirLabs API Function}
\usage{
airlabs_api_function(
  key,
  parameter_name = "parameter1",
  parameter_value = "value1",
  parameter_name2 = "parameter2",
  parameter_value2 = "value2"
)
}
\arguments{
\item{key}{The API key for accessing the AirLabs API.}

\item{parameter_name}{Name of the first parameter to include in the API request.}

\item{parameter_value}{Value of the first parameter.}

\item{parameter_name2}{Name of the second parameter to include in the API request.}

\item{parameter_value2}{Value of the second parameter.}
}
\value{
A data frame containing the response from the AirLabs API.
}
\description{
This function sends a request to the AirLabs API and returns the response as a data frame.
}
\examples{
# Retrieve data from the AirLabs API
data <- airlabs_api_function(key = "your_api_key", parameter_name = "parameter1", parameter_value = "value1",
                             parameter_name2 = "parameter2", parameter_value2 = "value2")
}
