% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airlines}
\alias{airlines}
\title{Airline lookup codes}
\format{
A data frame with airline lookup codes
}
\usage{
airlines
}
\description{
This dataset contains lookup codes for airlines.
}
\keyword{datasets}
