% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airports}
\alias{airports}
\title{Airport lookup codes}
\format{
A data frame with airport lookup codes
}
\usage{
airports
}
\description{
This dataset contains lookup codes for airports.
}
\keyword{datasets}
