% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_delays.R
\name{analyse_delays}
\alias{analyse_delays}
\title{Analyze delays in airlines}
\usage{
analyse_delays(directory, threshold)
}
\arguments{
\item{directory}{The directory containing the delay data files.}

\item{threshold}{The threshold for the delay data collection.}
}
\value{
A list containing three charts and a data table.
}
\description{
This function analyzes delay data for airlines based on a given directory
and threshold.
}
