% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delayed_percentage.R
\name{delayed_percentage}
\alias{delayed_percentage}
\title{Calculate the percentage of delayed flights}
\usage{
delayed_percentage(data, terminal_type)
}
\arguments{
\item{data}{A data frame containing flight data.}

\item{terminal_type}{A character string specifying the terminal type.
Must be either "dep" for departures or "arr" for arrivals.}
}
\value{
A data frame with the calculated percentages and counts for each category.
}
\description{
This function takes flight data and calculates the percentage of flights that are on time,
within 15 minutes of the scheduled time, and late. The calculations can be performed for
either departure or arrival flights based on the specified terminal type.
}
\examples{
delayed_percentage(data, "dep")
delayed_percentage(data, "arr")

}
