% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_live_flight_data.R
\docType{data}
\name{flight_type}
\alias{flight_type}
\title{Get live flight data for departures or arrivals}
\format{
An object of class \code{character} of length 1.
}
\usage{
flight_type
}
\arguments{
\item{flight_type}{A character string specifying the type of flight data to retrieve.
Must be either "departures" or "arrivals".}

\item{airport}{A character string specifying the airport code.
The flight data will be filtered based on this airport.}
}
\value{
A data frame containing the flight data.
}
\description{
This function retrieves live flight data for departures or arrivals from the Airlabs API
based on the specified flight type (departures or arrivals) and airport.
}
\examples{
get_flight_data("departures", "LAX")
get_flight_data("arrivals", "LAX")

}
\keyword{datasets}
