% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_delays.R
\name{get_delays}
\alias{get_delays}
\title{get_delays}
\usage{
get_delays(directory, threshold)
}
\arguments{
\item{directory}{The directory path where the CSV files are located.}

\item{threshold}{The threshold for filtering the files based on numeric values in their names.}
}
\value{
A list containing the average delay time and proportion of delayed flights.
}
\description{
This function filters CSV files in a directory based on a threshold, reads the selected files,
combines them into a single data frame, and calculates the average delay time and proportion
of flights that are delayed.
}
\examples{
filter_csv_files("path/to/csv/files", 5000)

}
