% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar_chart.R
\name{plot_bar_chart}
\alias{plot_bar_chart}
\title{Plot Bar Chart}
\usage{
plot_bar_chart(data, x, y, lab_x, lab_y, title)
}
\arguments{
\item{data}{The data frame containing the data to plot.}

\item{x}{The variable to use for the x-axis.}

\item{y}{The variable to use for the y-axis.}

\item{lab_x}{The label for the x-axis.}

\item{lab_y}{The label for the y-axis.}

\item{title}{The title of the plot.}
}
\value{
A bar chart plot.
}
\description{
This function creates a bar chart using the ggplot2 package.
}
\examples{
# Create a bar chart
plot_bar_chart(data = my_data, x = variable1, y = variable2,
               lab_x = "X-axis label", lab_y = "Y-axis label", title = "Bar Chart")
}
