% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_donut_chart.R
\name{plot_donut_chart}
\alias{plot_donut_chart}
\title{Create a donut chart to visualize performance metrics}
\usage{
plot_donut_chart(percentage_df, title)
}
\arguments{
\item{percentage_df}{A data frame containing the calculated percentages and counts for each category.
Must have columns "Category", "Percentage", and "Flights".}

\item{title}{A character string specifying the title of the chart.}
}
\description{
This function takes a data frame containing performance metrics and creates a donut chart
to visualize the percentages of flights that are on time, within 15 minutes, and late.
The chart is presented in a speedometer style.
}
\examples{
plot_donut_chart(percentage_df, "Departures")
plot_donut_chart(percentage_df, "Arrivals")

}
